refactor: split runner shared helpers
