+++ projects/04-llm-adapter-shadow/src/llm_adapter/runner_shared/__init__.py
@@ -1,0 +1,36 @@
+"""Shared runner helpers re-exported for backward compatibility."""
+from __future__ import annotations
+
+from .costs import estimate_cost
+from .logging import (
+    MetricsPath,
+    error_family,
+    log_provider_call,
+    log_provider_skipped,
+    log_run_metric,
+    resolve_event_logger,
+)
+from .rate_limiter import (
+    RateLimiter,
+    asyncio,
+    provider_model,
+    resolve_rate_limiter,
+    threading,
+    time,
+)
+
+__all__ = [
+    "MetricsPath",
+    "resolve_event_logger",
+    "error_family",
+    "estimate_cost",
+    "provider_model",
+    "log_provider_skipped",
+    "log_provider_call",
+    "log_run_metric",
+    "RateLimiter",
+    "resolve_rate_limiter",
+    "time",
+    "asyncio",
+    "threading",
+]
